sp_machine/requirements.txt
selenium
pypdf2
reportlab
openai
openpyxl
weasyprint
python-docx
pandas
flask
